pyproject.toml
[tool.poetry]
name = "chelmno_tv"
version = "0.1.0"
description = "fb and yt upload"
authors = ["Michal Bultrowicz <michal.bultrowicz@gmail.com>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
facebook-sdk = "^3.1.0"

[tool.poetry.dev-dependencies]
ptpython = "^3.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
